refactor: api.ts에서 interface 코드 다른 .ts 파일로 분할
